added all the things
